requirements.txt
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
python-jose
pymysql      # For MySQL
python-dotenv
# passlib
bcrypt==4.0.1
